requirements.txt
torch>=1.8.1
onnxruntime-gpu>=1.16.0
torchcrepe>=0.0.21
torchvision>=0.6.1
tqdm>=4.30
librosa>=0.9
opencv_python>=4.2.0
openai-whisper>=20230918
moviepy>=1.0.3
yt-dlp>=2023.9.24
gradio>=3.46.1
argparse>=1.4.0
wget>=3.2
audio-separator>=0.7.3
elevenlabs>=0.2.26
deep-translator>=1.11.4